+++ sync-for-expenses/src/sdk/models/operations/getlatestsync.ts
@@ -19,7 +19,7 @@ export class GetLatestSyncResponse extends SpeakeasyBase {
      * Success
      */
     @SpeakeasyMetadata()
-    companySyncStatus?: shared.CompanySyncStatus;
+    companySyncStatus?: Record<string, any>;
 
     /**
      * HTTP response content type for this operation