+++ sync-for-expenses/src/sdk/models/shared/pushoperationchange.ts
@@ -4,8 +4,7 @@
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { PushChangeType } from "./pushchangetype";
-import { PushOperationRef } from "./pushoperationref";
-import { Expose, Type } from "class-transformer";
+import { Expose } from "class-transformer";
 
 export class PushOperationChange extends SpeakeasyBase {
     /**
@@ -17,8 +16,7 @@ export class PushOperationChange extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     @Expose({ name: "recordRef" })
-    @Type(() => PushOperationRef)
-    recordRef?: PushOperationRef;
+    recordRef?: Record<string, any>;
 
     /**
      * Type of change being applied to record in third party platform.