+++ sync-for-expenses/src/sdk/models/shared/pushoptionproperty.ts
@@ -5,7 +5,6 @@
 import { objectToClass, SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { PushOptionChoice } from "./pushoptionchoice";
 import { PushOptionType } from "./pushoptiontype";
-import { PushValidationInfo } from "./pushvalidationinfo";
 import { Expose, Transform, Type } from "class-transformer";
 
 export class PushOptionProperty extends SpeakeasyBase {
@@ -58,6 +57,5 @@ export class PushOptionProperty extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     @Expose({ name: "validation" })
-    @Type(() => PushValidationInfo)
-    validation?: PushValidationInfo;
+    validation?: Record<string, any>;
 }