+++ sync-for-expenses/src/sdk/models/shared/validation.ts
@@ -3,20 +3,17 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { ValidationItem } from "./validationitem";
-import { Expose, Type } from "class-transformer";
+import { Expose } from "class-transformer";
 
 /**
  * A human-readable object describing validation decisions Codat has made when pushing data into the platform. If a push has failed because of validation errors, they will be detailed here.
  */
 export class Validation extends SpeakeasyBase {
-    @SpeakeasyMetadata({ elemType: ValidationItem })
+    @SpeakeasyMetadata()
     @Expose({ name: "errors" })
-    @Type(() => ValidationItem)
-    errors?: ValidationItem[];
+    errors?: Record<string, any>[];
 
-    @SpeakeasyMetadata({ elemType: ValidationItem })
+    @SpeakeasyMetadata()
     @Expose({ name: "warnings" })
-    @Type(() => ValidationItem)
-    warnings?: ValidationItem[];
+    warnings?: Record<string, any>[];
 }