+++ sync-for-expenses/src/sdk/sdk.ts
@@ -31,6 +31,7 @@ export type SDKProps = {
      * The security details required to authenticate the SDK
      */
     security?: shared.Security | (() => Promise<shared.Security>);
+
     /**
      * Allows overriding the default axios client used by the SDK
      */
@@ -58,9 +59,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "prealpha";
-    sdkVersion = "3.1.0";
-    genVersion = "2.146.1";
-    userAgent = "speakeasy-sdk/typescript 3.1.0 2.146.1 prealpha @codat/sync-for-expenses";
+    sdkVersion = "3.2.0";
+    genVersion = "2.150.0";
+    userAgent = "speakeasy-sdk/typescript 3.2.0 2.150.0 prealpha @codat/sync-for-expenses";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);